relegate all page logic to layout to allow Gantt chart and other project UI to persist
