+++ app/(dashboard)/[workspaceId]/[projectId]/[sprintId]/page.tsx
@@ -1,0 +1,3 @@
+export default function SprintPage() {
+  return null;
+}
